Split profile classes into separate modules
