+++ photutils/profiles/__init__.py
@@ -4,3 +4,5 @@ This subpackage contains tools for generating radial profiles.
 """
 
 from .core import *  # noqa: F401, F403
+from .curve_of_growth import *  # noqa: F401, F403
+from .radial_profile import *  # noqa: F401, F403
