+++ photutils/profiles/tests/test_curve_of_growth.py
@@ -10,8 +10,8 @@ from astropy.modeling.models import Gaussian2D
 from astropy.utils.exceptions import AstropyUserWarning
 from numpy.testing import assert_allclose, assert_equal
 
-from photutils.aperture import CircularAnnulus, CircularAperture
-from photutils.profiles import CurveOfGrowth, RadialProfile
+from photutils.aperture import CircularAperture
+from photutils.profiles import CurveOfGrowth
 from photutils.utils._optional_deps import HAS_MATPLOTLIB
 
 
@@ -193,66 +193,3 @@ def test_curve_of_growth_plot(profile_data):
                         radius_step, error=error << unit, mask=None)
     cg3.plot()
     cg3.plot_error()
-
-
-def test_radial_profile(profile_data):
-    xycen, data, _, _ = profile_data
-
-    min_radius = 0.0
-    max_radius = 35.0
-    radius_step = 1.0
-    rp1 = RadialProfile(data, xycen, min_radius, max_radius, radius_step,
-                        error=None, mask=None)
-
-    assert_equal(rp1.radius, np.arange(36))
-    assert rp1.area.shape == (36,)
-    assert rp1.profile.shape == (36,)
-    assert rp1.profile_error.shape == (0,)
-    assert rp1.area[0] > 0.0
-
-    assert len(rp1.apertures) == 36
-    assert isinstance(rp1.apertures[0], CircularAperture)
-    assert isinstance(rp1.apertures[1], CircularAnnulus)
-
-    min_radius = 0.7
-    max_radius = 35.0
-    radius_step = 1.0
-    rp2 = RadialProfile(data, xycen, min_radius, max_radius, radius_step,
-                        error=None, mask=None)
-    assert isinstance(rp2.apertures[0], CircularAnnulus)
-
-
-def test_radial_profile_unit(profile_data):
-    xycen, data, error, _ = profile_data
-
-    min_radius = 0.0
-    max_radius = 35.0
-    radius_step = 1.0
-    unit = u.Jy
-    rp1 = RadialProfile(data << unit, xycen, min_radius, max_radius,
-                        radius_step, error=error << unit, mask=None)
-    assert rp1.profile.unit == unit
-    assert rp1.profile_error.unit == unit
-
-    with pytest.raises(ValueError):
-        RadialProfile(data << unit, xycen, min_radius, max_radius, radius_step,
-                      error=error, mask=None)
-
-
-def test_radial_profile_error(profile_data):
-    xycen, data, error, _ = profile_data
-
-    min_radius = 0.0
-    max_radius = 35.0
-    radius_step = 1.0
-    rp1 = RadialProfile(data, xycen, min_radius, max_radius, radius_step,
-                        error=error, mask=None)
-
-    assert_equal(rp1.radius, np.arange(36))
-    assert rp1.area.shape == (36,)
-    assert rp1.profile.shape == (36,)
-    assert rp1.profile_error.shape == (36,)
-
-    assert len(rp1.apertures) == 36
-    assert isinstance(rp1.apertures[0], CircularAperture)
-    assert isinstance(rp1.apertures[1], CircularAnnulus)
